+++ Coding/ban_ve_xe_khach_web/src/main/java/com/cnwtt/ban_ve_xe_khach/api/VeXeResource.java
@@ -28,7 +28,7 @@ public class VeXeResource {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<VeXe> get(@PathVariable("id") Integer id) {
+    public ResponseEntity<VeXe> get(@PathVariable("id") Long id) {
         LOGGER.info("Call api get vexe: id[{}]", id);
         return ResponseEntity.ok(service.findById(id));
     }
@@ -40,7 +40,7 @@ public class VeXeResource {
     }
 
     @DeleteMapping("/delete/{id}")
-    public void delete(@PathVariable("id") Integer id) {
+    public void delete(@PathVariable("id") Long id) {
         LOGGER.info("Call api delete vexe : id [{}]", id);
         service.delete(id);
     }
