+++ Coding/ban_ve_xe_khach_web/src/main/java/com/cnwtt/ban_ve_xe_khach/repository/NhaXeRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface NhaXeRepository extends JpaRepository<NhaXe, Integer> {
+public interface NhaXeRepository extends JpaRepository<NhaXe, Long> {
     Page<NhaXe> findAll(Pageable pageable);
 
 }
